# Conflicts:
#	Gopkg.lock
#	Gopkg.toml
#	data/block/block.go
#	data/dataPool/nonceToHashCacher.go
#	data/dataPool/nonceToHashCacher_test.go
#	data/errors.go
#	data/interface.go
#	data/shardedData/shardedData.go
#	data/shardedData/shardedData_test.go
#	data/transaction/transaction.go
#	data/typeConverters/interface.go
#	data/typeConverters/uint64ByteSlice/bigEndianConverter.go
#	data/typeConverters/uint64ByteSlice/bigEndianConverter_test.go
#	marshal/capnpMarshalizer_test.go
#	p2p/memMessenger_test.go
#	p2p/netMessenger_test.go
#	p2p/topic.go
#	p2p/topic_test.go
#	process/block/export_test.go
#	process/block/interceptors.go
#	process/block/interceptors_test.go
#	process/errors.go
#	process/mock/shardedDataStub.go
#	process/mock/transactionInterceptorMock.go
#	process/transaction/export_test.go
#	process/transaction/interceptedTransaction.go
#	process/transaction/interceptedTransaction_test.go
#	process/transaction/txInterceptor.go
#	process/transaction/txInterceptor_test.go
#	storage/storageunit.go

+++ consensus/spos/consensus.go
@@ -161,7 +161,7 @@ func (cns *Consensus) GetLeader() (string, error) {
 		return "", errors.New("consensusGroup is empty")
 	}
 
-	index := cns.Chr.Round().Index() % len(cns.consensusGroup)
+	index := cns.Chr.Round().Index() % int32(len(cns.consensusGroup))
 	return cns.consensusGroup[index], nil
 }
